requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
markdown2==2.4.10
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
